requirements.txt
keras_vggface==0.6
mtcnn==0.1.0
numpy==1.21.4
opencv_python==4.5.5.64
Pillow==9.2.0
scikit_learn==1.1.1
streamlit==1.2.0
tensorflow==2.8.0
tqdm==4.62.3
Pillow==9.0.0
sklearn